requirements.txt
flask
pillow
pyheif #requires brew install libheif
